Устранение дребезга и разделение фильтров на два модуля
