+++ js/filters/filters-utils.ts
@@ -1,5 +1,4 @@
 import { Photo } from '../contracts/common';
-import { renderThumbnails } from '../posts/render-thumbnails';
 import { getRandomIdPack } from '../utils';
 
 const filters = document.querySelector('.img-filters')!;
@@ -7,9 +6,7 @@ const defaultButton = filters.querySelector('#filter-default')!;
 const randomButton = filters.querySelector('#filter-random')!;
 const discussedButton = filters.querySelector('#filter-discussed')!;
 
-function compareComments({ comments: commentsA }: Photo, { comments: commentsB }: Photo) {
-	return commentsB.length - commentsA.length;
-}
+const compareComments = ({ comments: commentsA }: Photo, { comments: commentsB }: Photo) => commentsB.length - commentsA.length;
 
 const defaultFilter = (photos: Photo[]) => photos;
 
@@ -42,20 +39,4 @@ const toggleActiveState = (event: Event) => {
 	targetButton.classList.add('img-filters__button--active');
 };
 
-const initFilters = (photos: Photo[]) => {
-	filters.classList.remove('img-filters--inactive');
-
-	const onFilterButtonClick = (evt: Event) => {
-		toggleActiveState(evt);
-		const currentElement = evt.target as HTMLButtonElement;
-		const currentFilterName = currentElement.getAttribute('id')!.split('-')[1];
-		const currentFilter = filterOptions[currentFilterName as keyof typeof filterOptions];
-		const currentPhotos = currentFilter(photos);
-		document.querySelectorAll('.picture').forEach((el) => el.remove());
-		renderThumbnails(currentPhotos);
-	};
-
-	[defaultButton, randomButton, discussedButton].forEach((button) => button.addEventListener('click', onFilterButtonClick));
-};
-
-export { initFilters };
+export { filterOptions, toggleActiveState, filters };
