+++ js/main.ts
@@ -2,12 +2,18 @@ import { renderThumbnails } from './posts/render-thumbnails';
 import { getData } from './fetch/api';
 import { showStatus } from './fetch/status';
 import './photo-upload/form';
-import { initFilters } from './filters/filters';
+import { setFilters } from './filters/set-filters';
+import { debounce } from './utils';
+
+const RERENDERER_DELAY = 500;
 
 getData()
 	.then((photos) => {
 		renderThumbnails(photos);
-		initFilters(photos);
+		setFilters(
+			debounce(renderThumbnails, RERENDERER_DELAY),
+			photos
+		);
 	})
 	.catch(() => {
 		showStatus('dataError');
